test(server): colocate status probe fixtures under unit/server (#1111)
Drop tests/helpers and regression re-exports; keep endpoint coverage in the required CI path only.
